+++ src/pages/Pools/Home/Favorites/index.tsx
@@ -7,7 +7,7 @@ import { useBondedPools } from 'contexts/Pools/BondedPools';
 import { usePoolsConfig } from 'contexts/Pools/PoolsConfig';
 import { useUi } from 'contexts/UI';
 import { CardWrapper } from 'library/Card/Wrappers';
-import { PoolList } from 'library/PoolList';
+import { PoolList } from 'library/PoolList/Default';
 import { useEffect, useState } from 'react';
 import { useTranslation } from 'react-i18next';
